Corrected typo in Definition 1.3
Corrected typo on page 4 in Definition 1.3. In the third condition of the definition of a Sample space it was stated:
"\item If $ A \in \mathcal{A} $ then $ \Omega\backslash \mathcal{A} \in \mathcal{A} $".
This should be: "\item If $A \in \mathcal{A}$ then $\Omega \setminus A \in \mathcal{A}$".